moved msal out of lib
