+++ app/api/msal/userImage/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { callMsGraph, getAccessToken } from '@/lib/msal';
+import { callMsGraph, getAccessToken } from '@/msal';
 
 export async function GET(req: NextRequest) {
   const { searchParams } = new URL(req.url);
