+++ msal/config.js
